requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
google-genai 
google-generativeai
dotenv
pydantic==2.8.2
gTTS==1.2.2
PyPDF2==3.0.1      # Para PDFs
python-docx==1.1.0 # Para .docx
pandas==2.2.2      # Para Excel/CSV
